Kommentek lekérdezéseinek kiszervezése
- game mapson belül a connections átnevezése paths-re
- paths lekérdezéseinek kiszervezése
