--- backend/api/gamemaps/connections/connections.service.js
@@ -1,9 +1,0 @@
-const database = require("#sql/database.js");
-
-async function getPointConnections(pointID) {
-    return await database.getConnectionsByPointId(pointID);
-}
-
-module.exports = {
-    getPointConnections
-};
