+++ backend/api/gamemaps/paths/__mocks__/paths.queries.js
@@ -1,0 +1,5 @@
+const mockPathQueries = {
+    getConnectionsByPointId: jest.fn()
+};
+
+module.exports = mockPathQueries;
