+++ backend/sql/__mocks__/database.js
@@ -19,14 +19,6 @@ const mockDatabase = {
     deleteImageById: jest.fn().mockResolvedValue(true),
     getPointImage: jest.fn(),
     insertConnection: jest.fn(),
-
-    getGameMapComments: jest.fn(),
-    getGameMapCommentCount: jest.fn(),
-    hasUserCommentedOnGameMap: jest.fn().mockResolvedValue(true),
-    getUserCommentOnGameMap: jest.fn(),
-    insertGameMapComment: jest.fn(),
-    updateUserCommentOnGameMap: jest.fn().mockResolvedValue(true),
-    deleteUserCommentOnGameMap: jest.fn().mockResolvedValue(true),
     getGameMapCoverImage: jest.fn(),
     updateGameMapCoverImage: jest.fn().mockResolvedValue(true),
     getTopScoresForGameMap: jest.fn(),
