+++ backend/sql/database.js
@@ -494,16 +494,6 @@ async function getPointImage(pointId) {
     return rows[0];
 }
 
-async function getConnectionsByPointId(pointId) {
-    const query = `
-        SELECT point_connections.connection_id, point_connections.start_point_id, point_connections.end_point_id
-        FROM point_connections
-        WHERE ? IN (point_connections.start_point_id, point_connections.end_point_id)
-    `;
-    const [rows] = await pool.execute(query, [pointId]);
-    return rows;
-}
-
 async function getMapImageIdByMapId(mapId) {
     const query = `
         SELECT map.image_id, images.filepath
@@ -659,86 +649,6 @@ async function updateGameMapDetails(connection, gameMapId, title, description) {
     return isIdUpdateSuccessful(result);
 }
 
-async function getGameMapComments(gameMapId, page) {
-    const safePage = Number.isInteger(Number(page)) && page > 0 ? page : 1;
-    const offset = (safePage - 1) * 50;
-    const query = `
-        SELECT 
-            COALESCE(users.username, 'Ismeretlen felhasználó') AS username,
-            game_maps_comments.rating,
-            game_maps_comments.comment_text,
-            game_maps_comments.created_at
-        FROM game_maps_comments
-            LEFT JOIN users ON (game_maps_comments.user_id = users.user_id)
-        WHERE game_maps_comments.game_maps_id = ?
-        ORDER BY game_maps_comments.created_at DESC
-        LIMIT 50 OFFSET ${offset}
-    `;
-    const [rows] = await pool.execute(query, [gameMapId]);
-    return rows;
-}
-
-async function getGameMapCommentCount(gameMapId) {
-    const query = `
-        SELECT 
-            COUNT(*) AS comment_count
-        FROM game_maps_comments
-        WHERE game_maps_comments.game_maps_id = ?
-    `;
-    const [rows] = await pool.execute(query, [gameMapId]);
-    return rows[0].comment_count;
-}
-
-async function hasUserCommentedOnGameMap(gameMapId, userId) {
-    const query = `
-        SELECT COUNT(*) AS comment_count
-        FROM game_maps_comments
-        WHERE game_maps_comments.game_maps_id = ?
-          AND game_maps_comments.user_id = ?
-    `;
-    const [rows] = await pool.execute(query, [gameMapId, userId]);
-    return rows[0].comment_count > 0;
-}
-
-async function insertGameMapComment(connection, gameMapId, userId, commentText, rating) {
-    const query = `
-        INSERT INTO game_maps_comments (game_maps_id, user_id, comment_text, rating)
-        VALUES (?, ?, ?, ?)
-    `;
-    const [result] = await connection.execute(query, [gameMapId, userId, commentText, rating]);
-    return result.insertId;
-}
-
-async function getUserCommentOnGameMap(gameMapId, userId) {
-    const query = `
-        SELECT game_maps_comments.comment_id, game_maps_comments.comment_text, game_maps_comments.rating, game_maps_comments.created_at
-        FROM game_maps_comments
-        WHERE game_maps_id = ? AND user_id = ?
-    `;
-    const [rows] = await pool.execute(query, [gameMapId, userId]);
-    return rows[0] || null;
-}
-
-async function updateUserCommentOnGameMap(connection, gameMapId, userId, commentText, rating) {
-    const query = `
-        UPDATE game_maps_comments
-        SET game_maps_comments.comment_text = ?, game_maps_comments.rating = ?
-        WHERE game_maps_comments.game_maps_id = ? AND game_maps_comments.user_id = ?
-    `;
-    const [result] = await connection.execute(query, [commentText, rating, gameMapId, userId]);
-
-    return isIdUpdateSuccessful(result);
-}
-
-async function deleteUserCommentOnGameMap(connection, gameMapId, userId) {
-    const query = `
-        DELETE FROM game_maps_comments
-        WHERE game_maps_comments.game_maps_id = ? AND game_maps_comments.user_id = ?
-    `;
-    const [result] = await connection.execute(query, [gameMapId, userId]);
-    return result.affectedRows == 1;
-}
-
 async function getAllImageIdsForGameMap(connection, gameMapId) {
     const query = `
         SELECT DISTINCT images.image_id
@@ -769,7 +679,6 @@ module.exports = {
     updateImagePath,
     getMapImage,
     getPointImage,
-    getConnectionsByPointId,
     checkUserOwnsGameMap,
     checkUserOwnsMap,
     checkUserOwnsPoint,
@@ -798,13 +707,6 @@ module.exports = {
     getGameMapCoverImage,
     updateGameMapCoverImage,
     updateGameMapDetails,
-    getGameMapComments,
-    getGameMapCommentCount,
-    hasUserCommentedOnGameMap,
-    insertGameMapComment,
-    getUserCommentOnGameMap,
-    updateUserCommentOnGameMap,
-    deleteUserCommentOnGameMap,
     getAllImageIdsForGameMap,
     deleteGameMapById,
     isIdUpdateSuccessful
